refactor curves directories
