+++ src/curves/interpolation/utils.rs
@@ -1,65 +1,8 @@
-
-use crate::dual::{FieldOps, MathFuncs, DualsOrF64};
-use crate::calendars::{CalType, Convention};
-use std::ops::Mul;
-use chrono::NaiveDateTime;
-use crate::curves::curve::NodesTimestamp;
-
-/// Interpolation 
-pub enum CurveInterpolator {
-    LogLinear(LogLinear),
-    Linear,
-    LinearIndex,
-    LinearZeroRate,
-    FlatForward,
-    FlatBackward,
-}
-
-pub struct LogLinear {
-    calendar: CalType,
-    convention: Convention,
-}
-
-/// Assigns methods for returning values from datetime indexed Curves.
-trait CurveInterpolation {
-    /// Get an item from the Curve expressed in its input form, i.e. discount factor or value.
-    fn interpolated_value(&self, nodes: &NodesTimestamp, date: &NaiveDateTime) -> DualsOrF64;
-
-    /// Get the left side node key index of the given datetime
-    fn node_index(&self, nodes: &NodesTimestamp, date_timestamp: &i64) -> usize {
-        // let timestamp = date.and_utc().timestamp();
-        index_left(&nodes.keys(), &date_timestamp, None)
-    }
-
-}
-
-impl CurveInterpolation for LogLinear {
-    fn interpolated_value(&self, nodes: &NodesTimestamp, date: &NaiveDateTime) -> DualsOrF64 {
-        let x = date.and_utc().timestamp();
-        let index = self.node_index(nodes, &x);
-//         let (x1, y1) = nodes.get_index(index);
-//         let (x2, y2) = nodes.get_index(index + 1_usize);
-//         log_linear_interp(*x1 as f64, y1, *x2 as f64, y2, x as f64)
-
-        match nodes {
-            NodesTimestamp::F64(m) => {
-                let (x1, y1) = m.get_index(index).unwrap();
-                let (x2, y2) = m.get_index(index + 1_usize).unwrap();
-                DualsOrF64::F64(log_linear_interp(*x1 as f64, y1, *x2 as f64, y2, x as f64))
-            }
-            NodesTimestamp::Dual(m) => {
-                let (x1, y1) = m.get_index(index).unwrap();
-                let (x2, y2) = m.get_index(index + 1_usize).unwrap();
-                DualsOrF64::Dual(log_linear_interp(*x1 as f64, y1, *x2 as f64, y2, x as f64))
-            }
-            NodesTimestamp::Dual2(m) => {
-                let (x1, y1) = m.get_index(index).unwrap();
-                let (x2, y2) = m.get_index(index + 1_usize).unwrap();
-                DualsOrF64::Dual2(log_linear_interp(*x1 as f64, y1, *x2 as f64, y2, x as f64))
-            }
-        }
-    }
-}
+use crate::dual::{FieldOps, MathFuncs};
+use std::{
+    ops::Mul,
+    cmp::{PartialEq, PartialOrd}
+};
 
 // pub(crate) fn linear_interp<T, U>(x1: &T, y1: &U, x2: &T, y2: &U, x: &T) -> U
 // where
@@ -127,31 +70,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use indexmap::IndexMap;
-    use crate::curves::curve::Nodes;
-    use crate::calendars::{NamedCal, ndt};
     use crate::dual::Dual;
-
-    fn nodes_timestamp_fixture() -> NodesTimestamp {
-        let nodes = Nodes::F64(IndexMap::from_iter(vec![
-            (ndt(2000, 1, 1), 1.0_f64),
-            (ndt(2001, 1, 1), 0.99_f64),
-            (ndt(2002, 1, 1), 0.98_f64),
-        ]));
-        NodesTimestamp::from(nodes)
-    }
-
-    #[test]
-    fn test_log_linear() {
-        let nts = nodes_timestamp_fixture();
-        let ll = LogLinear {
-            calendar: CalType::NamedCal(NamedCal::try_new("all").unwrap()),
-            convention: Convention::Act365F,
-        };
-        let result = ll.interpolated_value(&nts, &ndt(2000, 7, 1));
-        // expected = exp(0 + (182 / 366) * (ln(0.99) - ln(1.0)) = 0.995015
-        assert_eq!(result, DualsOrF64::F64(0.9950147597711371));
-    }
 
     #[test]
     fn index_left_() {
